+++ resources/hpxml-measures/HPXMLtoOpenStudio/resources/constructions.rb
@@ -1316,6 +1316,108 @@ class Constructions
     settings.setSimulationTimestep('Timestep')
   end
 
+  def self.apply_kiva_initial_temp(foundation, slab, weather, conditioned_zone,
+                                   sim_begin_month, sim_begin_day, sim_year,
+                                   foundation_walls_insulated, foundation_ceiling_insulated)
+    # Set Kiva foundation initial temperature
+
+    outdoor_temp = weather.data.MonthlyAvgDrybulbs[sim_begin_month - 1]
+
+    # Approximate indoor temperature
+    if conditioned_zone.thermostatSetpointDualSetpoint.is_initialized
+      # Building has HVAC system
+      setpoint_sch = conditioned_zone.thermostatSetpointDualSetpoint.get
+      sim_begin_date = OpenStudio::Date.new(OpenStudio::MonthOfYear.new(sim_begin_month), sim_begin_day, sim_year)
+
+      # Get heating/cooling setpoints for the simulation start
+      htg_setpoint_sch = setpoint_sch.heatingSetpointTemperatureSchedule.get
+      if htg_setpoint_sch.to_ScheduleRuleset.is_initialized
+        htg_day_sch = htg_setpoint_sch.to_ScheduleRuleset.get.getDaySchedules(sim_begin_date, sim_begin_date)[0]
+        heat_setpoint = UnitConversions.convert(htg_day_sch.values[0], 'C', 'F')
+      else
+        heat_setpoint = 78.0 # F, from ASHRAE 152 (avoids runtime impact from processing ScheduleFile CSV)
+      end
+      clg_setpoint_sch = setpoint_sch.coolingSetpointTemperatureSchedule.get
+      if clg_setpoint_sch.to_ScheduleRuleset.is_initialized
+        clg_day_sch = clg_setpoint_sch.to_ScheduleRuleset.get.getDaySchedules(sim_begin_date, sim_begin_date)[0]
+        cool_setpoint = UnitConversions.convert(clg_day_sch.values[0], 'C', 'F')
+      else
+        cool_setpoint = 68.0 # F, from ASHRAE 152 (avoids runtime impact from processing ScheduleFile CSV)
+      end
+
+      # Methodology adapted from https://github.com/NREL/EnergyPlus/blob/b18a2733c3131db808feac44bc278a14b05d8e1f/src/EnergyPlus/HeatBalanceKivaManager.cc#L303-L313
+      heat_balance_temp = UnitConversions.convert(10.0, 'C', 'F')
+      cool_balance_temp = UnitConversions.convert(15.0, 'C', 'F')
+      if outdoor_temp < heat_balance_temp
+        indoor_temp = heat_setpoint
+      elsif outdoor_temp > cool_balance_temp
+        indoor_temp = cool_setpoint
+      elsif cool_balance_temp == heat_balance_temp
+        indoor_temp = heat_balance_temp
+      else
+        weight = (cool_balance_temp - outdoor_temp) / (cool_balance_temp - heat_balance_temp)
+        indoor_temp = heat_setpoint * weight + cool_setpoint * (1.0 - weight)
+      end
+    else
+      # Building does not have HVAC system
+      indoor_temp = outdoor_temp
+    end
+
+    # Determine initial temperature
+    # For unconditioned spaces, this overrides EnergyPlus's built-in assumption of 22C (71.6F);
+    #   see https://github.com/NREL/EnergyPlus/blob/b18a2733c3131db808feac44bc278a14b05d8e1f/src/EnergyPlus/HeatBalanceKivaManager.cc#L257-L259
+    # For conditioned spaces, this avoids an E+ 22.2 bug; see https://github.com/NREL/EnergyPlus/issues/9692
+    if HPXML::conditioned_locations.include? slab.interior_adjacent_to
+      initial_temp = indoor_temp
+    else
+      # Space temperature assumptions from ASHRAE 152 - Duct Efficiency Calculations.xls, Zone temperatures
+      ground_temp = weather.data.GroundMonthlyTemps[sim_begin_month - 1]
+      if slab.interior_adjacent_to == HPXML::LocationBasementUnconditioned
+        if foundation_ceiling_insulated
+          # Insulated ceiling: 75% ground, 25% outdoor, 0% indoor
+          ground_weight, outdoor_weight, indoor_weight = 0.75, 0.25, 0.0
+        elsif foundation_walls_insulated
+          # Insulated walls: 50% ground, 0% outdoor, 50% indoor (case not in ASHRAE 152)
+          ground_weight, outdoor_weight, indoor_weight = 0.5, 0.0, 0.5
+        else
+          # Uninsulated: 50% ground, 20% outdoor, 30% indoor
+          ground_weight, outdoor_weight, indoor_weight = 0.5, 0.2, 0.3
+        end
+        initial_temp = outdoor_temp * outdoor_weight + ground_temp * ground_weight + indoor_weight * indoor_temp
+      elsif slab.interior_adjacent_to == HPXML::LocationCrawlspaceVented
+        if foundation_ceiling_insulated
+          # Insulated ceiling: 90% outdoor, 10% indoor
+          outdoor_weight, indoor_weight = 0.9, 0.1
+        elsif foundation_walls_insulated
+          # Insulated walls: 25% outdoor, 75% indoor (case not in ASHRAE 152)
+          outdoor_weight, indoor_weight = 0.25, 0.75
+        else
+          # Uninsulated: 50% outdoor, 50% indoor
+          outdoor_weight, indoor_weight = 0.5, 0.5
+        end
+        initial_temp = outdoor_temp * outdoor_weight + indoor_weight * indoor_temp
+      elsif slab.interior_adjacent_to == HPXML::LocationCrawlspaceUnvented
+        if foundation_ceiling_insulated
+          # Insulated ceiling: 85% outdoor, 15% indoor
+          outdoor_weight, indoor_weight = 0.85, 0.15
+        elsif foundation_walls_insulated
+          # Insulated walls: 25% outdoor, 75% indoor
+          outdoor_weight, indoor_weight = 0.25, 0.75
+        else
+          # Uninsulated: 40% outdoor, 60% indoor
+          outdoor_weight, indoor_weight = 0.4, 0.6
+        end
+        initial_temp = outdoor_temp * outdoor_weight + indoor_weight * indoor_temp
+      elsif slab.interior_adjacent_to == HPXML::LocationGarage
+        initial_temp = outdoor_temp + 11.0
+      else
+        fail "Unhandled space: #{slab.interior_adjacent_to}"
+      end
+    end
+
+    foundation.setInitialIndoorAirTemperature(UnitConversions.convert(initial_temp, 'F', 'C'))
+  end
+
   def self.create_insulation_material(model, name, rvalue)
     rigid_mat = BaseMaterial.InsulationRigid
     mat = OpenStudio::Model::StandardOpaqueMaterial.new(model)